feat: Add async write API
This adds a `tokio` feature that allows for async writes using
`tokio::io::AsyncWriteExt`. It also adds a `sync` feature, enabled by
default, that exposes the original sync API.

Code duplication between the two APIs is avoided using a combination of
[maybe-async] and the `write_bound` attribute macro defined in
`hepmc2-macros`. The former automatically removes `async`/`await`
keywords for the sync API and the latter is used to choose the correct
trait bounds based on which API is enabled.

[maybe-async]: https://crates.io/crates/maybe-async

Signed-off-by: Kiran Ostrolenk <kiran@ostrolenk.co.uk>